% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_control.R
\name{app_control}
\alias{app_control}
\title{app_control}
\usage{
app_control(config = NULL)
}
\arguments{
\item{config}{A list of config settings. If NULL, the function will open
config.yml for you.}
}
\value{
Sets the global variables run_flag and email_flag. These control if things run or not and allows staff to disable reports from App Control 1.
}
\description{
To run or not to run. That 'tis the question. Based on the IDEA app, App
Control 1. This is a "naughty" fuction. It exists for the side-effects. Used
to enable/disable code chunks and specific parts of the code.
}
