% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stratifications.R
\name{clean_stratifications}
\alias{clean_stratifications}
\title{clean_stratifications}
\usage{
clean_stratifications(s)
}
\arguments{
\item{s}{A vector of stratifications}
}
\value{
A vector of cleaned stratifications.
}
\description{
Removes things like /, \, :, ", ', <, > which would be invalid file names.
All such values are replaced with -, which is OK in a file name.
}
