% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_urls.R
\name{extract_urls}
\alias{extract_urls}
\title{extract_urls}
\usage{
extract_urls(text = NULL)
}
\arguments{
\item{text}{The text string you want to extract URLs from.}
}
\value{
A vector of URLs from the text.
}
\description{
Extract URLs from a string.
}
