% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stratifications.R
\name{get_stratifications}
\alias{get_stratifications}
\title{get_stratifications}
\usage{
get_stratifications(data = NULL, strat_col = NULL)
}
\arguments{
\item{data}{The data the stratifications are hiding in.}

\item{strat_col}{Which column they are hiding in.}
}
\value{
A vector of the stratifications. If there aren't any, "".
}
\description{
Returns the stratifications for a report.
}
