% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(
  file = "query.sql",
  folder = "sql",
  qry = NULL,
  config = NULL,
  verbose = FALSE,
  params = NULL
)
}
\arguments{
\item{file}{The file where the query lives.}

\item{folder}{The folder where the query lives. If validation.sql is in this
folder it will be run BEFORE the sql file.}

\item{qry}{A query to import, without a file.}

\item{verbose}{If TRUE, this function talks a lot. Useful for debugging
connection issues and not much else.}

\item{params}{A LIST of parameters to alter the query.}
}
\value{
A data frame with your data. If the query doesn't return anything,
you will receive an empty tibble.
}
\description{
Looks for "file" in "folder". If it exists, it attempts to run the query.
The user may also pass a query directly to import_data.
}
