% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_description.R
\name{read_description}
\alias{read_description}
\title{read_description}
\usage{
read_description(values = NULL)
}
\arguments{
\item{values}{A vector of the values you want to read from Description}
}
\value{
A list of packages to load via library.
}
\description{
Reads in values from the Description file.
}
