% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_config.R
\name{report_config}
\alias{report_config}
\title{report_config}
\usage{
report_config(config = NULL, report_todo = FALSE)
}
\arguments{
\item{config}{A configlist. If not provided, the function will open
config.yml itself.}
}
\value{
A tibble which is ready to be dropped into a report.
}
\description{
Uses the values in config.yml to provide report status information.
}
