% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{send_email}
\usage{
send_email(emails_to_send = NULL)
}
\arguments{
\item{emails_to_send}{A table of emails to send.}
}
\value{
Returns emails sent status for each record in emails_to_send
}
\description{
Sends the recipients in config.yml an email.
They get their report. We get paid. Everyone wins.
When run with the prod or rsconnect configs, it logs the sent email data to
the EDW (dev and prod).
}
