% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config_value.R
\name{set_config_value}
\alias{set_config_value}
\title{set_config_value}
\usage{
set_config_value(parameter, value, configuration = "default", preview = FALSE)
}
\arguments{
\item{parameter}{Which parameter do you want to set?}

\item{value}{What value do you want to set?}

\item{configuration}{Which named configuration from config.yml do you want to
use? Must be "default", "prod", or "rsconnect". The default is . . . default.}

\item{preview}{Do you want to just "see" what it would look like? This is for
interactive/testing purposes. Don't use this in a script. Seriously. It won't
do anything useful. The default is zero.}
}
\value{
None
}
\description{
Sets the values of EXISTING parameters. You have to be careful with this
thing, today, because it can only set EXISTING values. But, as long as they
exist, you can hack away.
}
\details{
There is a preview mode, so you can admire your handiwork.
}
