% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(file = "query.sql", folder = "sql", verbose = FALSE, params = NULL)
}
\arguments{
\item{file}{The file where the query lives.}

\item{folder}{The folder where the query lives. If validation.sql is in this
folder it will be run BEFORE the sql file.}

\item{verbose}{If TRUE, this function talks a lot. Useful for debugging
connection issues and not much else.}

\item{params}{A LIST of parameters to alter the query.}
}
\value{
A data frame with your data.
}
\description{
Looks for "file" in "folder". If it exists, it attempts to run the query
AFTER running validation.sql if it exists. If validation.sql exists, all rows
must return 1 (TRUE) before the report query in "file" will be run. If
validation.sql fails, the function will stop for the number of seconds
defined in wait_time before trying to rerun validation.sql.
}
