% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_con.R
\name{open_con}
\alias{open_con}
\title{open_con}
\usage{
open_con()
}
\value{
A connection object.
}
\description{
Uses config.yml and ~/.Renviron to open a database connection to the Acuitas
EDW. In your .Renviron file you must have the following edw_user and edw_pass
defined. In your config.yml file, you must have dsn_name defined for both the
default and rsconnect configurations.
}
\details{
Because this function uses config.yml and .Renviron, there are NO input
parameters to this function.
}
