% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config}
\alias{validate_config}
\title{validate_config}
\usage{
validate_config(file = "./config.yml")
}
\arguments{
\item{file}{The config.yml file you want to validate. You can specify any
file, but it defaults to ./config.yml.}
}
\description{
Validates that the config.yml is good.
}
